Adds mouseover UI features
When the user mouses over a portion of the main panel that has a
function associated with it, a new, mostly-transparent image (with
highlighing for the selected clickable) will be set as the background
image of the main panel. The actual background image has now been moved
to a new widget, that way we can use two background images at once.
The cursor will also change when you hover over a clickable.

Qt is silly, and wouldn't let me do what I wanted to do, so I used a
terribly dirty hack to get everything to work. In an ideal world, this
would be fixed and done correctly, but I probably won't have a chance
to do that, considering the bad luck I've already had with figuring out
Qt.